rename LayerStyles files to ThemeStyles
