+++ packages/app/src/containers/ThemeInspector.tsx
@@ -2,7 +2,7 @@ import { Selectors } from 'noya-state';
 import { memo, useMemo } from 'react';
 import { useSelector } from '../contexts/ApplicationStateContext';
 import SwatchInspector from './SwatchesInspector';
-import LayerStyleInspector from './LayerStyleInspector';
+import ThemeStyleInspector from './ThemeStyleInspector';
 
 export default memo(function ThemeInspector() {
   const tab = useSelector(Selectors.getCurrentComponentsTab);
@@ -12,7 +12,7 @@ export default memo(function ThemeInspector() {
       case 'swatches':
         return <SwatchInspector />;
       case 'layerStyles':
-        return <LayerStyleInspector />;
+        return <ThemeStyleInspector />;
     }
     return null;
   }, [tab]);
